finish stock data provider

+++ dataProvider/src/main/java/com/gergo/takacs/stock/dataprovider/PreferredStockEntity.java
@@ -13,7 +13,7 @@ public class PreferredStockEntity extends StockEntity {
 
 	/**
 	 */
-	private float fixedDividend;
+	private double fixedDividend;
 
 	/**
 	 */
